refactor: remove duplicate code from feed endpoints
